# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Webmot - Random Quote Generator
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Quote {
  id: number;
  text: string;
  author: string;
  category: string;
  categoryLabel: string;
  imageUrl: string;
  imageAlt: string;
}

export interface Category {
  slug: string;
  label: string;
  count: number;
}

export type ListQuotesParams = {
/**
 * Filter quotes by category slug
 */
category?: string;
};

export type GetRandomQuoteParams = {
/**
 * Restrict random selection to a specific category slug
 */
category?: string;
};