# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Webmot - Random Quote Generator
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  slug: string;
  label: string;
  count: number;
}