# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Webmot - Random Quote Generator
 * OpenAPI spec version: 0.1.0
 */

export interface Quote {
  id: number;
  text: string;
  author: string;
  category: string;
  categoryLabel: string;
  imageUrl: string;
  imageAlt: string;
}